+++ internal/aghos/os.go
@@ -5,7 +5,16 @@ import (
 	"fmt"
 	"os/exec"
 	"syscall"
+
+	"github.com/AdguardTeam/golibs/errors"
 )
+
+// ErrUnsupported is returned when the functionality is unsupported on the
+// current operating system.
+//
+// TODO(a.garipov): Make a structured error and use it everywhere instead of
+// a bunch of fmt.Errorf and all that.
+const ErrUnsupported errors.Error = "unsupported"
 
 // CanBindPrivilegedPorts checks if current process can bind to privileged
 // ports.
@@ -13,10 +22,11 @@ func CanBindPrivilegedPorts() (can bool, err error) {
 	return canBindPrivilegedPorts()
 }
 
-// SetRlimit sets user-specified limit of how many fd's we can use
-// https://github.com/AdguardTeam/AdGuardHome/internal/issues/659.
-func SetRlimit(val uint) {
-	setRlimit(val)
+// SetRlimit sets user-specified limit of how many fd's we can use.
+//
+// See https://github.com/AdguardTeam/AdGuardHome/internal/issues/659.
+func SetRlimit(val uint64) (err error) {
+	return setRlimit(val)
 }
 
 // HaveAdminRights checks if the current user has root (administrator) rights.